requirements.txt
flask
python-dotenv
google-generativeai
requests
gunicorn